fix: circular deps b/w configurable-enum-ordering.ts->configurable-enum.interface.ts
