+++ src/app/child-dev-project/attendance/model/attendance-status.ts
@@ -1,4 +1,5 @@
-import { ConfigurableEnumValue } from "../../../core/basic-datatypes/configurable-enum/configurable-enum.interface";
+import { ConfigurableEnumValue } from "app/core/basic-datatypes/configurable-enum/configurable-enum.types";
+
 
 /**
  * logical type of an attendance status, i.e. how it will be considered for statistics and analysis.
