+++ src/app/child-dev-project/children/demo-data-generators/educational-material/materials.ts
@@ -1,4 +1,5 @@
-import { ConfigurableEnumValue } from "../../../../core/basic-datatypes/configurable-enum/configurable-enum.interface";
+import { ConfigurableEnumValue } from "app/core/basic-datatypes/configurable-enum/configurable-enum.types";
+
 
 export const materials: ConfigurableEnumValue[] = [
   {
