+++ src/app/child-dev-project/children/model/genders.ts
@@ -1,4 +1,4 @@
-import { ConfigurableEnumValue } from "../../../core/basic-datatypes/configurable-enum/configurable-enum.interface";
+import { ConfigurableEnumValue } from "app/core/basic-datatypes/configurable-enum/configurable-enum.types";
 
 export const genders: ConfigurableEnumValue[] = [
   {
