+++ src/app/child-dev-project/notes/model/interaction-type.interface.ts
@@ -1,4 +1,5 @@
-import { ConfigurableEnumValue } from "../../../core/basic-datatypes/configurable-enum/configurable-enum.interface";
+import { ConfigurableEnumValue } from "app/core/basic-datatypes/configurable-enum/configurable-enum.types";
+
 
 /**
  * Additional properties defined in the "interaction-type" {@link ConfigurableEnumValue} values
