+++ src/app/core/basic-datatypes/configurable-enum/configurable-enum-datatype/configurable-enum.datatype.ts
@@ -1,8 +1,9 @@
-import { ConfigurableEnumValue } from "../configurable-enum.interface";
+
 import { EntitySchemaField } from "../../../entity/schema/entity-schema-field";
 import { ConfigurableEnumService } from "../configurable-enum.service";
 import { Injectable } from "@angular/core";
 import { DiscreteDatatype } from "../../discrete/discrete.datatype";
+import { ConfigurableEnumValue } from "../configurable-enum.types";
 
 @Injectable()
 export class ConfigurableEnumDatatype extends DiscreteDatatype<
