+++ src/app/core/basic-datatypes/configurable-enum/configurable-enum-directive/configurable-enum.directive.ts
@@ -1,6 +1,7 @@
 import { Directive, Input, TemplateRef, ViewContainerRef } from "@angular/core";
 import { ConfigurableEnumService } from "../configurable-enum.service";
-import { ConfigurableEnumValue } from "../configurable-enum.interface";
+import { ConfigurableEnumValue } from "../configurable-enum.types";
+
 
 /**
  * Enumerate over all {@link ConfigurableEnumConfig} values for the given enum config id.
