+++ src/app/core/basic-datatypes/configurable-enum/configurable-enum.service.ts
@@ -1,9 +1,10 @@
 import { Injectable } from "@angular/core";
 import { ConfigurableEnum } from "./configurable-enum";
 import { EntityMapperService } from "../../entity/entity-mapper/entity-mapper.service";
-import { ConfigurableEnumValue } from "./configurable-enum.interface";
+
 import { EntityAbility } from "../../permissions/ability/entity-ability";
 import { Entity } from "../../entity/model/entity";
+import { ConfigurableEnumValue } from "./configurable-enum.types";
 
 @Injectable({ providedIn: "root" })
 export class ConfigurableEnumService {
