+++ src/app/core/basic-datatypes/configurable-enum/configurable-enum.spec.ts
@@ -1,5 +1,6 @@
-import { ConfigurableEnumValue } from "./configurable-enum.interface";
+
 import { ConfigurableEnum } from "./configurable-enum";
+import { ConfigurableEnumValue } from "./configurable-enum.types";
 
 describe("ConfigurableEnum", () => {
   let sampleValues: ConfigurableEnumValue[];
