+++ src/app/core/basic-datatypes/configurable-enum/configurable-enum.types.ts
@@ -1,0 +1,14 @@
+
+
+export interface HasOrdinal {
+    _ordinal?: number;
+  }
+  
+  export interface ConfigurableEnumValue extends HasOrdinal {
+    id: string;
+    label: string;
+    color?: string;
+    isInvalidOption?: boolean;
+    style?: string;
+  }
+
