+++ src/app/core/basic-datatypes/configurable-enum/display-configurable-enum/display-configurable-enum.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit } from "@angular/core";
 import { ViewDirective } from "../../../entity/default-datatype/view.directive";
 import { DynamicComponent } from "../../../config/dynamic-components/dynamic-component.decorator";
-import { ConfigurableEnumValue } from "../configurable-enum.interface";
+
 import { NgClass, NgForOf, NgIf } from "@angular/common";
+import { ConfigurableEnumValue } from "../configurable-enum.types";
 
 /**
  * This component displays a {@link ConfigurableEnumValue} as text.
