+++ src/app/core/basic-datatypes/configurable-enum/edit-configurable-enum/edit-configurable-enum.component.ts
@@ -1,11 +1,12 @@
 import { Component, OnInit } from "@angular/core";
 import { EditComponent } from "../../../entity/default-datatype/edit-component";
-import { ConfigurableEnumValue } from "../configurable-enum.interface";
+
 import { DynamicComponent } from "../../../config/dynamic-components/dynamic-component.decorator";
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { ReactiveFormsModule } from "@angular/forms";
 import { MatSelectModule } from "@angular/material/select";
 import { EnumDropdownComponent } from "../enum-dropdown/enum-dropdown.component";
+import { ConfigurableEnumValue } from "../configurable-enum.types";
 
 @DynamicComponent("EditConfigurableEnum")
 @Component({
