+++ src/app/core/basic-datatypes/configurable-enum/enum-dropdown/enum-dropdown.component.ts
@@ -3,7 +3,7 @@ import { MatSelectModule } from "@angular/material/select";
 import { FormControl, ReactiveFormsModule } from "@angular/forms";
 import { ConfigurableEnumDirective } from "../configurable-enum-directive/configurable-enum.directive";
 import { NgForOf, NgIf } from "@angular/common";
-import { ConfigurableEnumValue } from "../configurable-enum.interface";
+
 import { BasicAutocompleteComponent } from "../../../common-components/basic-autocomplete/basic-autocomplete.component";
 import { ConfigurableEnumService } from "../configurable-enum.service";
 import { EntityMapperService } from "../../../entity/entity-mapper/entity-mapper.service";
@@ -16,6 +16,7 @@ import { ErrorHintComponent } from "../../../common-components/error-hint/error-
 import { MatButtonModule } from "@angular/material/button";
 import { ConfirmationDialogService } from "../../../common-components/confirmation-dialog/confirmation-dialog.service";
 import { OkButton } from "../../../common-components/confirmation-dialog/confirmation-dialog/confirmation-dialog.component";
+import { ConfigurableEnumValue } from "../configurable-enum.types";
 
 @Component({
   selector: "app-enum-dropdown",
