+++ src/app/core/common-components/entities-table/table-sort/table-sort.spec.ts
@@ -2,10 +2,10 @@ import { tableSort } from "./table-sort";
 import moment from "moment";
 import {
   ConfigurableEnumConfig,
-  ConfigurableEnumValue,
 } from "../../../basic-datatypes/configurable-enum/configurable-enum.interface";
 import { Entity } from "../../../entity/model/entity";
 import { Ordering } from "../../../basic-datatypes/configurable-enum/configurable-enum-ordering";
+import { ConfigurableEnumValue } from "app/core/basic-datatypes/configurable-enum/configurable-enum.types";
 
 describe("TableSort", () => {
   class E extends Entity {
